feat: Create skeleton routes and controllers for pins
