+++ routes/users.js
@@ -3,6 +3,7 @@ const router = express.Router()
 const userController = require('../controllers/userController')
 
 router.post('/signup', userController.signUp)
+
 router.post('/signin', userController.signIn)
 
 module.exports = router
